requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.26.165
botocore==1.29.165
certifi==2023.5.7
charset-normalizer==3.1.0
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
jmespath==1.0.1
outcome==1.2.0
packaging==23.1
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.6.1
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tqdm==4.65.0
trio==0.22.1
trio-websocket==0.10.3
urllib3==1.26.16
webdriver-manager==3.8.6
wsproto==1.2.0
html5lib
loguru
pdfkit